move tests to own directories
