+++ app/app/src/test/java/com/ar/enbaldeapp/models/ProductMust.java
@@ -1,4 +1,4 @@
-package com.ar.enbaldeapp;
+package com.ar.enbaldeapp.models;
 
 import static com.ar.enbaldeapp.support.Constants.PRODUCT_DESCRIPTION;
 import static com.ar.enbaldeapp.support.Constants.PRODUCT_ID;
