+++ app/app/src/test/java/com/ar/enbaldeapp/services/RegistrationApiServicesMust.java
@@ -1,4 +1,4 @@
-package com.ar.enbaldeapp;
+package com.ar.enbaldeapp.services;
 
 import static com.ar.enbaldeapp.support.Constants.ADDRESS;
 import static com.ar.enbaldeapp.support.Constants.EMAIL;
